add:查看app历史价格功能  完善修改app价格功能

+++ manager/src/main/java/com/celloud/manager/mapper/AppMapper.java
@@ -21,7 +21,7 @@ public interface AppMapper {
     int updateByPrimaryKeyWithBLOBs(App record);
 
     int updateByPrimaryKey(App record);
-    
+
     /**
      * 统计大客户下的app数量
      *
@@ -31,8 +31,9 @@ public interface AppMapper {
      * @author han
      * @date 2016年3月10日 下午2:10:51
      */
-    public int countApp(@Param("companyId") Integer companyId,@Param("offLine") int offLine);
-    
+    public int countApp(@Param("companyId") Integer companyId,
+            @Param("offLine") int offLine);
+
     /**
      * 统计大客户下的app运行次数
      *
@@ -44,8 +45,12 @@ public interface AppMapper {
      * @author han
      * @date 2016年3月11日 下午3:12:16
      */
-    public List<Map<String,Integer>> countAppRunNum(@Param("companyId") Integer companyId,@Param("offLine") int offLine,@Param("flag") Integer flag,@Param("period") int period,@Param("testAccountIds")String testAccountIds);
-    
+    public List<Map<String, Integer>> countAppRunNum(
+            @Param("companyId") Integer companyId,
+            @Param("offLine") int offLine, @Param("flag") Integer flag,
+            @Param("period") int period,
+            @Param("testAccountIds") String testAccountIds);
+
     /**
      * 统计大客户下的用户运行 app次数
      *
@@ -57,8 +62,12 @@ public interface AppMapper {
      * @author han
      * @date 2016年3月11日 下午3:22:33
      */
-    public List<Map<String,Integer>> countAppRunNumByUser(@Param("companyId") Integer companyId,@Param("offLine") int offLine,@Param("flag") Integer flag,@Param("period") int period,@Param("testAccountIds")String testAccountIds);
-    
+    public List<Map<String, Integer>> countAppRunNumByUser(
+            @Param("companyId") Integer companyId,
+            @Param("offLine") int offLine, @Param("flag") Integer flag,
+            @Param("period") int period,
+            @Param("testAccountIds") String testAccountIds);
+
     /**
      * 统计大客户下的app
      *
@@ -68,8 +77,9 @@ public interface AppMapper {
      * @author han
      * @date 2016年3月18日 下午2:44:02
      */
-    public List<App> getAppOfBigCustomer(@Param("companyId") Integer companyId,@Param("offLine") int offLine);
-    
+    public List<App> getAppOfBigCustomer(@Param("companyId") Integer companyId,
+            @Param("offLine") int offLine);
+
     /**
      * 大客户的APP价格列表
      * 
@@ -81,6 +91,5 @@ public interface AppMapper {
      * @date 2016年3月18日 下午4:47:11
      */
     public List<App> appPriceList(@Param("companyId") Integer companyId,
-            @Param("offLine") int offLine,
-            @Param("priceType") Integer priceType);
+            @Param("offLine") int offLine, @Param("priceType") Byte priceType);
 }
